Improve Line::containsPoint
Summary:
The old algorithm assumed an imaginary line with a fixed direction
this makes it so that the imaginary line is perpendicular to the line.

Test Plan: This can be tested by selecting a wire and hovering over a segment.

Reviewers: Joel

Reviewed By: Joel

Differential Revision: https://phabricator.simulton.com/D106

+++ qschematic/items/line.h
@@ -22,10 +22,10 @@ namespace QSchematic {
         bool isVertical() const;
         qreal lenght() const;
         QPointF midPoint() const;
-        bool containsPoint(const QPointF& point, unsigned tolerance = 0) const;
+        bool containsPoint(const QPointF& point, qreal tolerance = 0) const;
         QPointF pointOnLineClosestToPoint(const QPointF& point);
 
-        static bool containsPoint(const QLineF& line, const QPointF& point, unsigned tolerance = 0);
+        static bool containsPoint(const QLineF& line, const QPointF& point, qreal tolerance = 0);
 
     private:
         QPointF _p1;
